Provide token fixes

+++ token/main/main.go
@@ -8,8 +8,8 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"log"
-	//"os"
-	//"os/exec"
+	"os"
+	"os/exec"
 	"token/config"
 	pb "token/generated"
 	"token/server"
@@ -18,7 +18,7 @@ import (
 
 func main() {
 	cfg, _ := config.LoadConfig(".")
-	/*if cfg.Debug {
+	if cfg.Debug {
 		if err := os.RemoveAll("./generated"); err != nil {
 			log.Fatal(err)
 		}
@@ -28,7 +28,7 @@ func main() {
 		if err := exec.Command("protoc", "--go_out=generated", "--go-grpc_out=generated", "--proto_path=proto", "proto/*.proto").Run(); err != nil {
 			log.Fatal(err)
 		}
-	}*/
+	}
 	redisClient := redis.NewClient(&redis.Options{Addr: fmt.Sprintf("%s:%s", cfg.RedisHostname, cfg.RedisPort)})
 	if _, err := redisClient.Ping(context.Background()).Result(); err != nil {
 		log.Fatal(err)
@@ -48,7 +48,6 @@ func main() {
 		}
 		return nil
 	})
-	log.Println(cfg.Debug)
 	server.Server{Address: cfg.ServerAddress}.Launch(func(server *grpc.Server) {
 		pb.RegisterTokenServiceServer(server, accountService)
 	}, authInterceptor)
